Multitasking: pass a lock of the condition checker to wait_on_event
Now there are two versions of for queuing the current thread :  `wait_on_queue` and `wait`. `wait_on_queue` which takes a template function to check if condition is met, and a lock to be acquired and released before sleeping. `wait` in other hand is unconditioned waiting so it doesn't need a checker function nor a lock.
There are no need for looping on wait_on_event anymore, since the loop is embedded in it, and it is guaranteed, that only one thread is woken up and it satisfies the condition.